+++ include/types.hpp
@@ -4,7 +4,7 @@
 
 enum class Var_Types {
   INT,
-  FLOAT,
+  DOUBLE,
   BOOL,
   STRING,
   UNKNOWN
@@ -14,8 +14,8 @@ inline std::string var_type_to_string(Var_Types type) {
   switch (type) {
     case Var_Types::INT:
       return "int";
-    case Var_Types::FLOAT:
-      return "float";
+    case Var_Types::DOUBLE:
+      return "double";
     case Var_Types::BOOL:
       return "bool";
     case Var_Types::STRING:
@@ -30,8 +30,8 @@ inline std::string var_type_to_string(Var_Types type) {
 inline Var_Types string_to_var_type(const std::string& type) {
   if (type == "int") {
     return Var_Types::INT;
-  } else if (type == "float") {
-    return Var_Types::FLOAT;
+  } else if (type == "double") {
+    return Var_Types::DOUBLE;
   } else if (type == "bool") {
     return Var_Types::BOOL;
   } else if (type == "string") {
@@ -44,7 +44,7 @@ inline Var_Types string_to_var_type(const std::string& type) {
 enum class Func_Types {
   VOID,
   INT,
-  FLOAT,
+  DOUBLE,
   BOOL,
   STRING,
   UNKNOWN
@@ -56,8 +56,8 @@ inline std::string func_type_to_string(Func_Types type) {
       return "void";
     case Func_Types::INT:
       return "int";
-    case Func_Types::FLOAT:
-      return "float";
+    case Func_Types::DOUBLE:
+      return "double";
     case Func_Types::BOOL:
       return "bool";
     case Func_Types::STRING:
@@ -74,8 +74,8 @@ inline Func_Types string_to_func_type(const std::string& type) {
     return Func_Types::VOID;
   } else if (type == "int") {
     return Func_Types::INT;
-  } else if (type == "float") {
-    return Func_Types::FLOAT;
+  } else if (type == "double") {
+    return Func_Types::DOUBLE;
   } else if (type == "bool") {
     return Func_Types::BOOL;
   } else if (type == "string") {
